refactor(fmt): rename groovy-formatter to fmt with gvy.fmt package
- Renamed groovy-formatter/ directory to fmt/

- Updated module references in settings.gradle.kts

- Changed package from com.github.albertocavalcante.groovyformatter to com.github.albertocavalcante.gvy.fmt

- Updated all imports in groovy-lsp and tests modules
