+++ groovy-lsp/src/main/kotlin/com/github/albertocavalcante/groovylsp/cli/FormatCommand.kt
@@ -7,7 +7,7 @@ import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.multiple
 import com.github.ajalt.clikt.parameters.types.file
 import com.github.ajalt.mordant.terminal.Terminal
-import com.github.albertocavalcante.groovyformatter.OpenRewriteFormatter
+import com.github.albertocavalcante.gvy.fmt.OpenRewriteFormatter
 import java.io.File
 
 /**
